Change for comments

+++ news/views.py
@@ -30,30 +30,21 @@ def news_details(request, slug):
             comment = comment_form.save(commit=False)
             comment.post = post
             comment.save()
-            if request.headers.get('X-Requested-With') == 'XMLHttpRequest':
-                return JsonResponse({
-                    'author_name': comment.author_name,
-                    'text': comment.text,
-                    'pub_date': comment.pub_date.strftime('%Y-%m-%d %H:%M')
-                })
-            else:
-                messages.success(request, 'Your comment has been added.')
-                return redirect('news_details', slug=slug)
+            messages.success(request, 'Your comment has been added.')
+            return redirect('news_details', slug=slug)
         else:
-            if request.headers.get('X-Requested-With') == 'XMLHttpRequest':
-                # Respond with form errors if the form is invalid
-                return JsonResponse({'error': comment_form.errors.as_json()}, status=400)
-            else:
-                messages.error(request, 'There was an error with your comment.')
-                return redirect('news_details', slug=slug)
+            messages.error(request, 'There was an error with your comment.')
+            return redirect('news_details', slug=slug)
 
     comment_form = CommentForm()
-    return render(request, 'news_details.html', {
+    template = 'news_details.html'
+    context = {
         'post': post,
         'comments': comments,
         'comment_form': comment_form,
         'post_form': post_form,
-    })
+    }
+    return render(request, template, context)
 
 
 @login_required
@@ -71,12 +62,17 @@ def create_news(request):
 @login_required
 def edit_news(request, slug):
     post = get_object_or_404(News_Post, slug=slug)
+    if post.author != request.user:
+        messages.error(request, 'Access denied, this is not your news post.')
+        return redirect('news_details', slug=slug)
+
     if request.method == 'POST':
         form = NewsPostForm(request.POST, request.FILES, instance=post, prefix='post')
         if form.is_valid():
             form.save()
             messages.success(request, 'Post updated successfully.')
             return redirect('news_details', slug=slug)
+        messages.error(request, 'Error: please try again')
 
     return redirect('news_details', slug=slug)
 
@@ -85,7 +81,12 @@ def edit_news(request, slug):
 @login_required
 def delete_news(request, slug):
     post = get_object_or_404(News_Post, slug=slug)
+    if post.author != request.user:
+        messages.error(request, 'Access denied, this is not your news post.')
+        return redirect('news_details', slug=slug)
+
     post.delete()
+    messages.success(request, 'New post successfully deleted!')
     return redirect('news_list')
 
 
@@ -93,6 +94,10 @@ def delete_news(request, slug):
 @login_required
 def delete_comment(request, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admins can perform this action')
+        return redirect('news_details', slug=comment.post.slug)
+
     comment.delete()
     messages.success(request, "Comment deleted successfully.")
     return redirect('news_details', slug=comment.post.slug)
